refactor(proxy): extract initialize-params resolution into ProxyCore
The 220-line Xcode chat client-version sub-domain (UserDefaults
IDEChat*Version scanning, alias normalization, TOML merge) lived inside
RuntimeCoordinator+Initialization with every function duplicated as a
static plus a pure-forwarding instance wrapper. It is pure
environment/config logic, so it now lives once in ProxyCore as
InitializeHandshakeParams; the duplicate defaultInitializeParams literal
in DocumentationProviderManager is gone too, so the documentation
provider handshake can no longer drift from the main one.
