+++ Sources/ProxySession/Runtime/RuntimeCoordinator+Initialization.swift
@@ -383,7 +383,9 @@ extension RuntimeCoordinator {
         }
     }
     func makeInternalInitializeRequest(id: Int64) -> [String: Any] {
-        let mergedParams = resolvedInitializeParams().mapValues(\.foundationObject)
+        let mergedParams = InitializeHandshakeParams.resolved(
+            initializeParamsOverride: initializeParamsOverride
+        ).mapValues(\.foundationObject)
 
         return [
             "jsonrpc": "2.0",
@@ -391,216 +393,5 @@ extension RuntimeCoordinator {
             "method": "initialize",
             "params": mergedParams,
         ]
-    }
-
-    func resolvedInitializeParams() -> [String: JSONValue] {
-        Self.resolvedInitializeParams(initializeParamsOverride: initializeParamsOverride)
-    }
-
-    package static func resolvedInitializeParams(config: ProxyConfig) -> [String: JSONValue] {
-        let override = ProxyFileConfigLoader.loadInitializeParamsOverride(
-            configPath: config.configPath,
-            logger: ProxyLogging.make("config")
-        )
-        return resolvedInitializeParams(initializeParamsOverride: override)
-    }
-
-    package static func resolvedInitializeParams(
-        initializeParamsOverride: [String: JSONValue]?
-    ) -> [String: JSONValue] {
-        let mergedParams = ProxyFileConfigLoader.mergeJSONObjects(
-            defaultInitializeParams(),
-            overriding: initializeParamsOverride ?? [:]
-        )
-        guard hasExplicitClientVersionOverride(initializeParamsOverride: initializeParamsOverride) == false else {
-            return mergedParams
-        }
-        return applyingAutomaticClientVersion(to: mergedParams)
-    }
-
-    func defaultInitializeParams() -> [String: JSONValue] {
-        Self.defaultInitializeParams()
-    }
-
-    package static func defaultInitializeParams() -> [String: JSONValue] {
-        [
-            "protocolVersion": .string("2025-03-26"),
-            "capabilities": .object([:]),
-            "clientInfo": .object([
-                "name": .string(defaultProxyClientName()),
-                "version": .string(defaultProxyClientVersion()),
-            ]),
-        ]
-    }
-
-    func hasExplicitClientVersionOverride() -> Bool {
-        Self.hasExplicitClientVersionOverride(initializeParamsOverride: initializeParamsOverride)
-    }
-
-    package static func hasExplicitClientVersionOverride(
-        initializeParamsOverride: [String: JSONValue]?
-    ) -> Bool {
-        guard case .object(let clientInfo)? = initializeParamsOverride?["clientInfo"] else {
-            return false
-        }
-        return clientInfo["version"] != nil
-    }
-
-    func applyingAutomaticClientVersion(to params: [String: JSONValue]) -> [String: JSONValue] {
-        Self.applyingAutomaticClientVersion(to: params)
-    }
-
-    package static func applyingAutomaticClientVersion(to params: [String: JSONValue]) -> [String: JSONValue] {
-        guard case .object(var clientInfo)? = params["clientInfo"],
-              case .string(let clientName)? = clientInfo["name"] else {
-            return params
-        }
-
-        guard let resolvedVersion = xcodeChatClientVersion(for: clientName) else {
-            return params
-        }
-
-        clientInfo["version"] = .string(resolvedVersion)
-        var updated = params
-        updated["clientInfo"] = .object(clientInfo)
-        return updated
-    }
-
-    func defaultClientVersion(for clientName: String) -> String {
-        Self.defaultClientVersion(for: clientName)
-    }
-
-    package static func defaultClientVersion(for clientName: String) -> String {
-        xcodeChatClientVersion(for: clientName) ?? defaultProxyClientVersion()
-    }
-
-    func defaultProxyClientName() -> String {
-        Self.defaultProxyClientName()
-    }
-
-    package static func defaultProxyClientName() -> String {
-        "XcodeMCPKit"
-    }
-
-    func defaultProxyClientVersion() -> String {
-        Self.defaultProxyClientVersion()
-    }
-
-    package static func defaultProxyClientVersion() -> String {
-        "dev"
-    }
-
-    func xcodeChatClientVersion(for clientName: String) -> String? {
-        Self.xcodeChatClientVersion(for: clientName)
-    }
-
-    package static func xcodeChatClientVersion(for clientName: String) -> String? {
-        let defaults = UserDefaults(suiteName: "com.apple.dt.Xcode")?.dictionaryRepresentation() ?? [:]
-        return xcodeChatClientVersion(for: clientName, defaults: defaults)
-    }
-
-    func xcodeChatClientVersion(for clientName: String, defaults: [String: Any]) -> String? {
-        Self.xcodeChatClientVersion(for: clientName, defaults: defaults)
-    }
-
-    package static func xcodeChatClientVersion(for clientName: String, defaults: [String: Any]) -> String? {
-        let normalizedName = normalizedChatClientName(clientName)
-        guard !normalizedName.isEmpty else { return nil }
-
-        var exactMatches: [(stem: String, version: String)] = []
-        var aliasMatches: [(stem: String, version: String)] = []
-
-        for (key, value) in defaults {
-            guard key.hasPrefix("IDEChat"), key.hasSuffix("Version") else { continue }
-            guard let raw = value as? String, let version = xcodeChatVersionValue(from: raw) else {
-                continue
-            }
-
-            let stem = String(
-                key
-                    .dropFirst("IDEChat".count)
-                    .dropLast("Version".count)
-            )
-
-            let normalizedStem = normalizedChatClientName(stem)
-            if normalizedStem == normalizedName {
-                exactMatches.append((stem, version))
-                continue
-            }
-
-            if chatClientAliases(forVersionStem: stem).contains(normalizedName) {
-                aliasMatches.append((stem, version))
-            }
-        }
-
-        let orderedExactMatches = exactMatches.sorted { lhs, rhs in
-            lhs.stem.localizedStandardCompare(rhs.stem) == .orderedAscending
-        }
-        if let match = orderedExactMatches.first {
-            return match.version
-        }
-
-        let orderedAliasMatches = aliasMatches.sorted { lhs, rhs in
-            lhs.stem.localizedStandardCompare(rhs.stem) == .orderedAscending
-        }
-        return orderedAliasMatches.first?.version
-    }
-
-    func xcodeChatVersionValue(forDefaultsKey defaultsKey: String) -> String? {
-        Self.xcodeChatVersionValue(forDefaultsKey: defaultsKey)
-    }
-
-    package static func xcodeChatVersionValue(forDefaultsKey defaultsKey: String) -> String? {
-        guard let raw = UserDefaults(suiteName: "com.apple.dt.Xcode")?.string(forKey: defaultsKey) else {
-            return nil
-        }
-        return xcodeChatVersionValue(from: raw)
-    }
-
-    func xcodeChatVersionValue(from raw: String) -> String? {
-        Self.xcodeChatVersionValue(from: raw)
-    }
-
-    package static func xcodeChatVersionValue(from raw: String) -> String? {
-        guard
-            let data = raw.data(using: .utf8),
-            let object = try? JSONSerialization.jsonObject(with: data, options: []) as? [String: Any],
-            let version = object["version"] as? String,
-            !version.isEmpty
-        else {
-            return nil
-        }
-        return version
-    }
-
-    func chatClientAliases(forVersionStem stem: String) -> Set<String> {
-        Self.chatClientAliases(forVersionStem: stem)
-    }
-
-    package static func chatClientAliases(forVersionStem stem: String) -> Set<String> {
-        var aliases: Set<String> = []
-        let normalizedStem = normalizedChatClientName(stem)
-        if !normalizedStem.isEmpty {
-            aliases.insert(normalizedStem)
-        }
-
-        if stem.hasSuffix("Code") {
-            let baseStem = String(stem.dropLast("Code".count))
-            let normalizedBaseStem = normalizedChatClientName(baseStem)
-            if !normalizedBaseStem.isEmpty {
-                aliases.insert(normalizedBaseStem)
-            }
-        }
-
-        return aliases
-    }
-
-    func normalizedChatClientName(_ name: String) -> String {
-        Self.normalizedChatClientName(name)
-    }
-
-    package static func normalizedChatClientName(_ name: String) -> String {
-        let scalars = name.unicodeScalars.filter(CharacterSet.alphanumerics.contains)
-        return String(String.UnicodeScalarView(scalars)).lowercased()
     }
 }
